move file to appropriate folder
