+++ src/app.tsx
@@ -2,15 +2,11 @@ import 'src/global.css';
 
 import { Router } from 'src/routes/sections';
 
-import { useScrollToTop } from 'src/hooks/use-scroll-to-top';
-
 import { ThemeProvider } from 'src/theme/theme-provider';
 
 // ----------------------------------------------------------------------
 
 export default function App() {
-  useScrollToTop();
-
   return (
     <ThemeProvider>
       <Router />
