+++ src/sections/patient/table-no-data.tsx
@@ -1,6 +1,6 @@
 import type { TableRowProps } from '@mui/material/TableRow';
 import { Box, TableRow, TableCell, Typography, Button } from '@mui/material';
-import useFetchPatients from '../patient/view/useFetchPatients';
+import useFetchPatients from '../../hooks/useFetchPatients';
 
 type EmptyTableProps = TableRowProps & {
   searchQuery: string;
@@ -11,8 +11,10 @@ type EmptyTableProps = TableRowProps & {
 const EmptyTable = ({ searchQuery, isTableEmpty, handleOpen }: EmptyTableProps) => {
   const { error, refetch } = useFetchPatients();
   const tableCopy = isTableEmpty
-    ? "There are no patient records currently. Click on the button below to create one!"
-    : error ? `There was an error loading the data. Please try again!` : `No results found for '${searchQuery}'. Try typing a different entry.`;
+    ? 'There are no patient records currently. Click on the button below to create one!'
+    : error
+      ? `There was an error loading the data. Please try again!`
+      : `No results found for '${searchQuery}'. Try typing a different entry.`;
   return (
     <TableRow>
       <TableCell align="center" colSpan={7}>
@@ -24,5 +26,5 @@ const EmptyTable = ({ searchQuery, isTableEmpty, handleOpen }: EmptyTableProps) 
       </TableCell>
     </TableRow>
   );
-}
-export default EmptyTable
+};
+export default EmptyTable;
